Move to using declarative byte offsets when parsing
The addition of the _extract_byte_data() function, coupled to the byte
offset values in byte_offsets.py, dramatically reduces the amount of
repeated logic.
It also separates out the "what does each byte mean?" from the program
logic, which makes reviewing most of the chunk type values easier.
Having the byte offsets in a separate file required the utility
functions in a separate file to avoid circular dependencies.

Also in this commit is a move to parse the master chunks and all their
associated container and boma sub-chunks in one go. Now the for-loop in
musicdb_to_json needs only to consider the master chunks, and we
implicitly validate the number of subsections they claim to contain.